quine sink for proper parallel ingest. PPM observation processing in child actors to improve parallelism. Quine write time measurements

+++ src/main/scala/com/galois/adapt/AdaptConfig.scala
@@ -76,7 +76,7 @@ object AdaptConfig extends Utils {
     systemname: String,
     quitonerror: Boolean,
     logfile: String,
-    lmdbgigabytes: Long = 4
+    lmdbgigabytes: Int = 4
   )
 
   case class EnvironmentConfig(
